lib/power.go: Adjust import

+++ lib/power.go
@@ -1,7 +1,7 @@
 package lib
 
-import math
+import "math"
 
 func Power(a int, b int) int {
-	return math.Pow(a, b)
+	return math.Pow(float64(a), float64(b))
 }
